Finished project ready to pack and deploy

+++ routes/ideas.js
@@ -45,6 +45,10 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => {
     try {
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
         const updatedIdea = await Idea.findByIdAndUpdate(
             req.params.id,
             {
@@ -55,7 +59,12 @@ router.put('/:id', async (req, res) => {
             }, 
             { new: true }
         );
-        res.json({success: true, data: updatedIdea})        
+        return res.json({success: true, data: updatedIdea})        
+        }
+
+        // Username does not match
+        res.status(403).json({ success: false, error: 'Unauthorized User for Updating - Go Away!'})   
+
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong'})        
@@ -66,8 +75,17 @@ router.put('/:id', async (req, res) => {
 // Delete idea
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({ success: 'true', data: {} });
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({ success: 'true', data: {} });
+        }
+
+        // Usernames dont match
+        res.status(403).json({ success: false, error: 'Unauthorized User - Go Away!'})   
+
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong'})
